Expose required rolling parameter `window`, doc improvement
I exposed the required parameter `window` because without it, rolling will raise an exception, and made it a required parameter. I also added typehints that are backwards compatible to Python 3.7 as well as do not affect the runtime library requirements of the module (still only require numpy and pandas). This is done by using TYPE_CHECKING guard and using annotation-based typehints which allow backwards compatibility to python 3.7. I also added to the documentation that the `window` and `min_periods`, when integer, are the number of periods used in each rolling estimate (rather than explaining it is because of using lagged prices, .shift(1), etc)